+++ packages/builder-tools/connect-utils/vite-plugins/dynamic-base.ts
@@ -133,6 +133,6 @@ export function connectDynamicBasePlugin(
   };
 }
 
-function escapeForRegExp(s: string): string {
+export function escapeForRegExp(s: string): string {
   return s.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
 }