+++ packages/reactor-browser/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 6.2.2-dev.11 (2026-07-22)
+
+### 🚀 Features
+
+- **connect:** optimize available packages listing in settings package manager ([9f823e59a](https://github.com/powerhouse-inc/powerhouse/commit/9f823e59a))
+- **connect:** dynamic packages can contribute to the PWA configuration ([da867e21d](https://github.com/powerhouse-inc/powerhouse/commit/da867e21d))
+- **connect:** implement file handling for PWA, enabling OS-level document imports ([57c4a5e57](https://github.com/powerhouse-inc/powerhouse/commit/57c4a5e57))
+
+### ❤️ Thank You
+
+- Yasiel Cabrera @YasielCabrera
+
+## 6.2.2-dev.10 (2026-07-21)
+
+This was a version bump only for @powerhousedao/reactor-browser to align it with other projects, there were no code changes.
+
 ## 6.2.2-dev.9 (2026-07-20)
 
 This was a version bump only for @powerhousedao/reactor-browser to align it with other projects, there were no code changes.